requirements.txt
Django
django-filter
djangorestframework
djangorestframework-simplejwt
django-cors-headers